Use d3 without d3pie for pie chart.
This is necessary because we want to have better support for dynamic
updating of data (d3pie wants to redraw the chart every time). And d3pie
is very pretty but imo some of the prettiness is unnecessary, e.g. I
think it's okay to make a pie chart without showing labels and
percentages statically, taking advantage of our interactive medium to
sidestep prettification issues altogether.