Very slightly better svg:title instead of tooltip.
Made an abortive attempt to use d3-tip but it is not well-suited to pie
chart data (since the tooltip positioning is designed to look good on
rectangular shapes, so looks a bit random on arcs). It will probably be
easiest just to write a few mouse event listeners to create/move/remove
a div element.