Use 'brushend' event for better performance.
The 'brush' event is triggered on mousemove, so there are a lot of them
when dragging the brush around, and we don't want to re-filter the table
every time. This gives noticeably smoother brushing on my phone and in
firefox. (For some reason chromium was fast enough that I didn't notice
this mistake earlier.)