Create tooltips and decorate data points.
The idea is to make it easier to follow what's happening on the chart.
It's a bit messy right now because the appearance of tooltips can be
triggered in two different ways, which was done to try to support
devices without :hover and also allow more interactive :hover tooltips
where possible. But the messiness of the implementation is probably not
worth the result, so in future this will probably be changed a bit.