Refine structure and now actually compute sums.
Remove the <dl> and put its contents into <tfoot> because that looks
nicer and makes better logical sense (only problem might be if the table
gets super long, the user has to scroll to see the totals, which
actually are probably what the user first wants to see anyway).

Add formatters so the table is a bit less ugly. Now with proper units
and precision, and without breaking sort order, yay.

Also fix a couple of small bugs.