Handle 401 response from reporting api.
Make the assumption that 401 means token has expired. This is not
necessarily true, but should be okay if we assume the users are not
being sneaky. (If a user manually set a token in web storage, which the
reporting api rejected, the user would be told that the session expired
even though the session technically never started, whatever.)