Create "Load" report.
This is meant to correspond directly to "Recurrent Work Load Utilisation
by Node", taking a different approach than in the "Utilisation" report
which combined lots of reports in a confusing interface.

This report shows a subset of what's in the Utilisation report.
Eventually the Utilisation report will be replaced by several different
reports, each showing a different subset, hopefully to be more
user-friendly.

Also this changeset begins the transition to using NVD3, rather than
custom d3 code, for charts.