Add historical_usage table and chart.
The new table just reconstructs historical data from the other tables
already in stats_schema.sql. Previously this was being done on the
client side, but that was too slow, since it required iterating over the
whole instance table. This approach gives an approximate overview of
the resource usage each day; more precise data will be available in
other more precise reports.

The historical_usage table is created with the procedure
create_historical_usage, and there is no update_historical_usage because
the idea is that a cron job inserts one new row each day. But right now
there is no cron job and this is a bit dodgy and needs a bit of
finishing.

At some point in the future a column historical_usage.volume_storage
should be added to record data from the volume table. (The report spec
includes "additional allocated storage" as something to be shown.)