Add line chart showing project historical usage.
This reconstructs history from the instance table. Unlike the
utilisation report, it only fetches data one project (or, in future,
institution) at a time, reducing the loading time.

Using custom chart instead of nvd3 lineWithFocusChart. First it was
implemented using nvd3, but that chart has certain limitations which
made it too awkward to use. (Namely zooming and panning get restricted
to regions bounded by data points, preventing selection of arbitrary
regions, in particular making it not possible to view a single month
unless data points happen to lie on the month's boundary.) The custom
chart code needs a bit of cleaning up, but its current state suffices to
show a pretty picture, which is nice.