Remove obsolete "Utilisation" report.
This report has been superseded by the "Load" and "Project" reports,
which are meant to be easier to use and which run faster. One data table
was lost in this transition, but maybe it will be reinstated later,
depending on how useful the Powers That Be think it was.

This also allowed cleanup of some hacky old code for weird fetch query
keys (live_instance and last_updated), which are now replaced by
filtered queries or smarter code, respectively.