Rework login code, to fix ugly bug.
If an error occurred during authentication, an exception would be
thrown, which would not be caught anywhere. This would then cause the
"return false" to be skipped, so the form would be submitted, sending a
GET request ?username=admin&password=secret&url=whatever to the view
layer (not even to the auth url supplied). Now errors should all be
handled correctly, and this should not happen again. Hooray!