Serialise chart generation and calling.
Ensure that charts are "generated" (initialised) before trying to draw
them. This is not the most efficient approach, since it waits for all
charts to be ready before doing anything, but it is the simplest.

Previously, if fetching the data happened very fast (e.g. already in
cache) the code might try to render a chart before nvd3 had actually
finished setting it up, giving an error and breaking the report.